fix: spearbit audit. finding #2 (#3916)
## Summary
There is no check to ensure `ecrecover()` does not return zero. As such,
any invalid signature is treated as valid and returned as the zero
address.

## Impact Explanation
The impact of this could potentially be quite severe. Anyone could forge
any data package and push a fictitious update to a price feed that could
result in liquidations, improper fills, and other disastrous behavior.

## Likelihood Explanation
The likelihood of this happening is presumably low.

The code comments for `RedstoneConsumerBase.getAuthorisedSignerIndex`
clearly state:

```solidity
   * @dev This function must be implemented by the child consumer contract.
   * It should return a unique index for a given signer address if the signer
   * is authorised, otherwise it should revert
```

The integrator would have to overlook this or make a mistake in order
for this to have an impact. However, there is also no guarantee how the
`SignatureLib` will be used in the future which increases the chance of
unexpected behavior.

## Recommendation
Check the return value of `ecrecover` and revert on address zero.
Alternatively, consider replacing the library with Open Zeppelin's
SignatureChecker.sol.